refactor(bridge): extract AIMessageHandler
